subscriber-api/requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Settings
pydantic-settings>=2.1.0

# Database (shared with admin backend)
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0

# Redis
redis>=5.0.0

# Auth
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Validation
email-validator>=2.1.0

# S3/MinIO
boto3>=1.34.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
aiosqlite>=0.20.0
fakeredis>=2.21.0

# Linting
ruff>=0.2.0
